+++ src/langserver/toolsGetALAppContentRequest.ts
@@ -1,0 +1,10 @@
+export class ToolsGetALAppContentRequest {
+    appPath: string;
+    filePath: string;
+
+    constructor(newAppPath: string, newFilePath: string) {
+        this.appPath = newAppPath;
+        this.filePath = newFilePath;        
+    }
+
+}