+++ src/langserver/toolsGetALAppContentResponse.ts
@@ -1,0 +1,3 @@
+export class ToolsGetALAppContentResponse {
+    source: string | undefined;
+}