+++ src/langserver/toolsGetLibrarySymbolLocationRequest.ts
@@ -1,0 +1,10 @@
+export class ToolsGetLibrarySymbolLocationRequest {
+    libraryId: number;
+    path: number[];
+
+    constructor(newLibraryId: number, newPath: number[]) {
+        this.libraryId = newLibraryId;
+        this.path = newPath;
+    }
+
+}