+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Change Log
 
+## 3.0.5
+ - Issue #194 - Object definition not available - load definition from app file if ShowMyCode is set true
+ - Issue #237 - Support Workspace Trust
+ - New "azALDevTools.showExtensionLog" command to open extension log file
+
+ Thank you
+  - fvet for reporting issue #194
+  - lramos15 for reporting issue #237
+
 ## 3.0.4
  - Both Windows and Mac language servers use .net 5 now
  - Issue #233 - Use table field caption changed by the table extension